📦 chore(ramda-extension.ts): add new utility functions containsAny and between to extend Ramda library
The `containsAny` function checks if any items from the first array are present in the second array. It returns `true` if any items are found, otherwise `false`.

The `between` function checks if a number (`val`) is within a range defined by two numbers (`a` and `b`). It returns `true` if the number is within the range, otherwise `false`.