✨ feat(socket.ts): add new route for streaming data
This commit adds a new file `socket.ts` which contains a route for streaming data. The route is a GET request to `/stream` and returns a response with a header of `Content-Type` set to `application/octet-stream` and sends a `Uint16Array` of length 10 as the response body. This route is tagged with `stream` and has a summary of "Get an user info".