feat: Add Dockerfile, .dockerignore, .gcloudignore, exclude
Add Dockerfile for building and running the application in Docker container.
Add .dockerignore to exclude unnecessary files and directories from the Docker build context.
Add .gcloudignore to specify files and directories to be excluded when deploying to Google Cloud.
Add 'exclude' file to list specific files and directories to be ignored.